laser scanner
The laser scanner is based on the integration of computer science, optics and optics, mechanical engineering and industrial automation and refers to the applications of image processing. The laser scanner is designed for inspection, quality control of objects with smooth surfaces (metal and paper plates) and 3D scanning of parts, with high speed, high accuracy, 24-hour operation and repeatability of measurement.
With the help of this technology, the dimensions of the object in the image (axis perpendicular to the two-dimensional image) can be calculated and in fact, a three-dimensional scan of an object can be obtained. In this technology, a laser is used for this purpose because the line of light created by the laser has a very high contrast with other colors in the image.